Extract metallic instrumentation environment helpers into workspace crate
