+++ crates/metallic_env/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "metallic_env"
+version = "0.1.0"
+edition = "2024"
+description = "Environment variable helpers shared across Metallic instrumentation."
+license = "MIT OR Apache-2.0"
+publish = false
+
+[lib]
+path = "src/lib.rs"
+
+[dependencies]
