+++ src/metallic/instrument/mod.rs
@@ -1,7 +1,6 @@
 //! Unified instrumentation module scaffolding the upcoming metrics system.
 
 pub mod config;
-pub mod environment;
 pub mod event;
 pub mod exporters;
 pub mod gpu_profiler;
