+++ src/metallic/instrument/prelude.rs
@@ -1,11 +1,12 @@
 //! Convenience re-exports for instrumentation consumers.
 
 pub use crate::metallic::instrument::config::{AppConfig, AppConfigError};
-pub use crate::metallic::instrument::environment::{EnvVar, EnvVarGuard, Environment, InstrumentEnvVar};
 pub use crate::metallic::instrument::event::MetricEvent;
 pub use crate::metallic::instrument::exporters::{ChannelExporter, ConsoleExporter, JsonlExporter};
 pub use crate::metallic::instrument::recorder::{EnrichedMetricEvent, MetricExporter, MetricsLayer};
 pub use crate::record_metric;
+
+pub use metallic_env::{EnvVar, EnvVarGuard, Environment, InstrumentEnvVar};
 
 pub use chrono::{DateTime, Utc};
 pub use serde_json;
